% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_to_gma.R
\name{calc_bilateral_trade_df_gma}
\alias{calc_bilateral_trade_df_gma}
\title{Calculates bilateral trade data frame associated with Global Market Assumption}
\usage{
calc_bilateral_trade_df_gma(
  .tidy_iea_df,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  country = IEATools::iea_cols$country,
  provenience = "Provenience",
  matnames = "matnames",
  Share_Exports_By_Product = "Share_Exports_By_Product"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the associated bilateral trade data, using the Global Market Assumption, needs to be calculated.}

\item{year, method, energy_type, last_stage}{See \code{IEATools::iea_cols}.}

\item{country}{The name of the column in the output data frame that receives the flow,
i.e. the importing country for a given flow.
Default is \code{IEATools::iea_cols$country}.}

\item{provenience}{The name of the column in the output data frame that originates the flow,
i.e. the exporting country for a given flow.
Default is "Provenience".}

\item{matnames}{The column name for matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{Share_Exports_By_Product}{The name of the column containing the share of exports by country, for each product.
Default is "Share_Exports_By_Product".}
}
\value{
A data frame that represents the bilateral trade data associated to the input \code{.tidy_iea_df},
using the Global Market Assumption.
}
\description{
The function calculates the bilateral trade data associated to a \code{.tidy_iea_df}, using the Global Market Assumption.
}
\details{
A column specifying matrix name for each flows needs to be added before, most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
calc_bilateral_trade_df_gma() \%>\%
print()
}
