% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_helpers.R
\name{calc_global_exports}
\alias{calc_global_exports}
\title{Calculates global exports by product}
\usage{
calc_global_exports(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  exports = IEATools::interface_industries$exports,
  Global_Exports_By_Product = "Global_Exports_By_Product"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which global exports by product need to be calculated.}

\item{flow, method, energy_type, last_stage, year, ledger_side, flow_aggregation_point, product, e_dot}{See \code{IEATools::iea_cols}.}

\item{exports}{The name of exports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$exports}.}

\item{Global_Exports_By_Product}{The name of the column containing global exports by product.}
}
\value{
A data frame reporting global exports by product.
}
\description{
The function calculates global exports by product.
}
\examples{
tidy_AB_data \%>\%
calc_global_exports() \%>\%
print()
}
