% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_helpers.R
\name{calc_global_production_by_product}
\alias{calc_global_production_by_product}
\title{Calculates global production by product}
\usage{
calc_global_production_by_product(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  V_matrix = IEATools::psut_cols$V,
  R_matrix = IEATools::psut_cols$R,
  imports = IEATools::interface_industries$imports,
  Global_Production_By_Product = "Global_Production_By_Product"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which global production needs to be calculated.}

\item{country, ledger_side, flow_aggregation_point, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name for matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{V_matrix}{The name of the V matrix.
Default is \code{IEATools::psut_cols$V}.}

\item{R_matrix}{The name of the R matrix.
Default is \code{IEATools::psut_cols$R}.}

\item{imports}{The name of imports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$imports}.}

\item{Global_Production_By_Product}{The name of the new column returning global production by product.}
}
\value{
A data frame representing global production by product, for each year.
}
\description{
This function calculates global production by product. Included flows are flows belonging to either the
R or V matrices, excluding import flows.
}
\details{
Note: the function needs to have a column indicating matrix names added first, most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
calc_global_production_by_product() \%>\%
print()
}
