% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_helpers.R
\name{calc_imports_by_product}
\alias{calc_imports_by_product}
\title{Calculates total imports by product, for each country, for each year.}
\usage{
calc_imports_by_product(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  imports = IEATools::interface_industries$imports
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which total imports need to be calculated.}

\item{flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{imports}{The name of the column containing the quantity (in energy terms) of imported products.
Default is \code{IEATools::interface_industries$imports}.}
}
\value{
A data frame returning the value of imported energy by product, for each country and year.
}
\description{
The function calculates total imports by product, for each country, for each year.
}
\examples{
tidy_AB_data \%>\%
calc_imports_by_product() \%>\%
print()
}
