% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_helpers.R
\name{calc_share_imports_by_products}
\alias{calc_share_imports_by_products}
\title{Calculates the shares of imports to total consumption by product, for each country and year.}
\usage{
calc_share_imports_by_products(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  imports = IEATools::interface_industries$imports,
  matnames = "matnames",
  Total_Consumption_By_Product = "Total_Consumption_By_Product",
  Share_Imports_By_Product = "Share_Imports_By_Product"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the shares of imports over final consumption, by produt, need to be calculated.}

\item{flow, country, method, energy_type, last_stage, ledger_side, flow_aggregation_point, year, product, unit, e_dot}{See \code{IEATools::iea_cols}.}

\item{imports}{The name of the column containing the quantity (in energy terms) of imported products.
Default is \code{IEATools::interface_industries$imports}.}

\item{matnames}{The column name for matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{Total_Consumption_By_Product}{The name of the new column returning total consumption by product.}

\item{Share_Imports_By_Product}{The name of the new column returning the share of imports over total consumption.}
}
\value{
A data frame returning the share of imports over final consumption, for each product.
}
\description{
The function calculates the shares of imports to total consumption by product, for each country and year.
}
\details{
Total consumption is calculated using the \code{calc_total_consumption_by_product()} function.

Note: the function needs to have a column indicating matrix names added first, most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
calc_share_imports_by_products() \%>\%
print()
}
