% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_helpers.R
\name{calc_total_consumption_by_product}
\alias{calc_total_consumption_by_product}
\title{Calculates total consumption by product, for each country and year}
\usage{
calc_total_consumption_by_product(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  exports = IEATools::interface_industries$exports,
  matnames = "matnames",
  Y_matrix = "Y",
  U_feed_matrix = "U_feed",
  U_EIOU_matrix = "U_EIOU",
  Epsilon_matrix = "Epsilon",
  Total_Consumption_By_Product = "Total_Consumption_By_Product"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} to which the function should be applied.}

\item{flow, country, method, energy_type, last_stage, year, product, unit, e_dot}{See \code{IEATools::iea_cols}.}

\item{exports}{The name of exports flows in the input data.
Default is \code{IEATools::interface_industries$exports}.}

\item{matnames}{The column name for matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{Y_matrix}{The name of the final demand Y matrix.
Default is \code{IEATools::psut_cols$Y}.}

\item{U_feed_matrix}{The name of the U_feed matrix.
Default is \code{IEATools::psut_cols$U_feed}.}

\item{U_EIOU_matrix}{The name of the U_eiou matrix.
Default is \code{IEATools::psut_cols$U_eiou}.}

\item{Epsilon_matrix}{The name of the Epsilon matrix.
Default is \code{IEATools::psut_cols$epsilon}.}

\item{Total_Consumption_By_Product}{The name of the new column reporting total consumption by product.}
}
\value{
A data frame with total consumption by product, for each country and year.
}
\description{
The function calculates total consumption by product, for each country and year.
}
\details{
Flows included in the calculation of the total consumption by product are flows belonging to either the Y,
U_feed, or U_eiou matrices. Flows belonging to the Y matrix but representing exports, are excluded.
In addition, flows belonging to the Epsilon matrix and akin to a final demand (i.e. E.dot > 0) are also included, provided they do not represent exports.
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
calc_total_consumption_by_product() \%>\%
print()
}
