% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{convert_to_net_trade}
\alias{convert_to_net_trade}
\title{Converts data to net trade data}
\usage{
convert_to_net_trade(
  .tidy_iea_df,
  imports = IEATools::interface_industries$imports,
  exports = IEATools::interface_industries$exports,
  country = IEATools::iea_cols$country,
  e_dot = IEATools::iea_cols$e_dot,
  flow = IEATools::iea_cols$flow,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  product = IEATools::iea_cols$product,
  .net_imports = "Net_Imports"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which trade flows need to be converted to net trade flows.}

\item{imports}{The name of the Imports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$imports}.}

\item{exports}{The name of the Exports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$exports}.}

\item{country, e_dot, flow, year, ledger_side, flow_aggregation_point, product}{See \code{IEATools::iea_cols}.}

\item{.net_imports}{A temporary column that calculates net imports as the difference between imports and exports.
Default is "Net_Imports".}
}
\value{
A \code{.tidy_iea_df} for which trade flows are converted to net trade.
}
\description{
This function enables to convert a \code{.tidy_iea_df} to
a new \code{.tidy_iea_df} in which trade flows are reported as net flows.
So, if a given region or country both imports and exports a given product,
the function will report only the difference between imports and exports,
depending on which flow is dominant. See Details for uses.
}
\details{
The function may be use after aggregating regions with the
\code{IEATools::aggregate_regions()} function, or may be applied to IEA data for
aggregated regions (such as World) for which both imports and exports are reported
for a given product.
}
\examples{
# In this example, we gather all flows for countries A and B
# in a new "AB" region, for which both imports and exports are reported for some products.
tidy_AB_data \%>\%
dplyr::mutate(Country = "AB") \%>\%
dplyr::filter(stringr::str_detect(Flow, "Imports") | stringr::str_detect(Flow, "Exports")) \%>\%
print()
# After running the function, only either imports or exports are reported for each product.
tidy_AB_data \%>\%
dplyr::mutate(Country = "AB") \%>\%
convert_to_net_trade() \%>\%
dplyr::filter(stringr::str_detect(Flow, "Imports") | stringr::str_detect(Flow, "Exports")) \%>\%
print()
}
