% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{exports_to_balancing}
\alias{exports_to_balancing}
\title{Relocates export flows in the Balancing matrix}
\usage{
exports_to_balancing(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  e_dot = IEATools::iea_cols$e_dot,
  exports = IEATools::interface_industries$exports,
  balancing_matrix = "B"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which export flows need to be sent to the Balancing matrix.}

\item{flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{exports}{The string that identifies export flows.
Default is \code{IEATools::interface_industries$exports}.}

\item{balancing_matrix}{The name of the Balancing matrix.
Default is "B".}
}
\value{
A \code{.tidy_iea_df} with Stock changes flows relocated in the Balancing matrix.
}
\description{
Relocates export flows in the Balancing matrix
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
dplyr::filter(stringr::str_detect(Flow, "Exports")) \%>\%
exports_to_balancing() \%>\%
print()
}
