% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_to_dta.R
\name{find_list_dta_observations}
\alias{find_list_dta_observations}
\title{Find list of observations for which Domestic Technology Assumption is applicable}
\usage{
find_list_dta_observations(
  .tidy_iea_df,
  products_to_look_for = IEATools::products,
  requirement_matrices_list = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou,
    IEATools::psut_cols$U_feed),
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  imports = IEATools::interface_industries$imports,
  matnames = IEATools::mat_meta_cols$matnames,
  y_matrix = IEATools::psut_cols$Y,
  u_eiou_matrix = IEATools::psut_cols$U_eiou,
  u_feed_matrix = IEATools::psut_cols$U_feed,
  v_matrix = IEATools::psut_cols$V,
  r_matrix = IEATools::psut_cols$R
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the list of observations is needed.}

\item{products_to_look_for}{The list of products that need to be looked for as consumed products.
Default is \code{IEATools::products}.}

\item{requirement_matrices_list}{The list of matrices where the \code{products_to_look_for} are looked for.
Default is \code{c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou, IEATools::psut_cols$U_feed)}.}

\item{country, method, energy_type, last_stage, year, product, flow}{See \code{IEATools::iea_cols}.}

\item{imports}{The name of imports flows in the \code{.tidy_iea_df}.
Defaut is \code{IEATools::interface_industries$imports}.}

\item{matnames}{The column name for matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{y_matrix}{The name of the Y matrix.
Default is \code{IEATools::psut_cols$Y}.}

\item{u_eiou_matrix}{The name of the U_EIOU matrix.
Default is \code{IEATools::psut_cols$U_EIOU}.}

\item{u_feed_matrix}{The name of the U_feed matrix.
Default is \code{IEATools::psut_cols$U_feed}.}

\item{v_matrix}{The name of the V matrix.
Default is \code{IEATools::psut_cols$V}.}

\item{r_matrix}{The name of the R matrix.
Default is \code{IEATools::psut_cols$R}.}
}
\value{
A list of observations for which the Domestic Technology Assumption is applicable.
Each observation is defined as Country_Method_Energy.type_Last.stage_Year.
}
\description{
This function returns a list of observations for which the Domestic Technology Assumption is applicable,
in the format Country_Method_Energy.type_Last.stage_Year. See details for the process.
}
\details{
Strictly speaking, the Domestic Technology Assumption can only be formulated when at least one unit of each consumed product (be it in the U_feed, U_eiou, or Y matrices)
is produced domestically - i.e., appears at least once in the V matrix in a non-importing flow. The default run of the function is set up to return only those observations
that fulfil such a requirement.

However, one may want to be more flexible about the formulation of the Domestic Technology Assumption. Indeed, nothing prevents from formulating the Domestic
Technology Assumption in cases that do not fulfil the aforementioned conditions, but in that case, any demand for a given product that is not produced anywhere
(be it direct or indirect demand) in the domestic economy will not be translated into any upstream demand, which is an important caveat.

The \code{products_to_look_for} and \code{requirement_matrices_list} offer additional flexibility, so that only products passed as argument to the \code{products_to_look_for} argument
and matrices passed as arguments to the \code{requirement_matrices_list} will act as constraining elements. Indeed, the new condition for formulation of the Domestic Technology
Assumption becomes that all products passed as \code{products_to_look_for} argument and used one of the \code{requirement_matrices_list}, must be produced at least once in the domestic
conversion chain (i.e. must appear in V in a non-importing flow).

A column specifying matrix name for each flows needs to be added before, most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
find_list_dta_observations()
}
