% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{international_bunkers_to_balancing}
\alias{international_bunkers_to_balancing}
\title{Moves international bunkers flows to the Balancing matrix}
\usage{
international_bunkers_to_balancing(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  e_dot = IEATools::iea_cols$e_dot,
  ledger_side = IEATools::iea_cols$ledger_side,
  international_marine_bunkers = "International marine bunkers",
  international_aviation_bunkers = "International aviation bunkers",
  balancing_matrix = "B"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which Stock changes flows need to be sent to the Balancing matrix.}

\item{flow, e_dot, ledger_side}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{international_marine_bunkers}{The name of the international marine bunkers flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$international_marine_bunkers}.}

\item{international_aviation_bunkers}{The name of the international aviation bunkers flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$international_aviation_bunkers}.}

\item{balancing_matrix}{The name of the Balancing matrix.
Default is "B".}
}
\value{
The new \code{.tidy_iea_df} with international bunkers ascribed to the Balancing matrix.
}
\description{
This function moves international bunkers flows to the Balancing matrix.
The Balancing matrix is akin to an additional final demand matrix,
meaning that flows akin to final demand (i.e. where stocks increase) will be positive,
while flows akin to supply (i.e. where stocks decrease) will be negative.
}
\details{
See the Balancing matrix vignette for more information.
Note: one needs to add the column containing matrices names first,
most likely using the \code{IEATools::add_psut_matnames()} function.
}
