% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{international_bunkers_to_epsilon}
\alias{international_bunkers_to_epsilon}
\title{Moves international bunkers flows to the Epsilon matrix}
\usage{
international_bunkers_to_epsilon(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  e_dot = IEATools::iea_cols$e_dot,
 
    international_marine_bunkers = IEATools::interface_industries$international_marine_bunkers,
 
    international_aviation_bunkers = IEATools::interface_industries$international_aviation_bunkers,
  epsilon = IEATools::psut_cols$epsilon
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which Stock changes flows need to be sent to the Epsilon matrix.}

\item{flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{international_marine_bunkers}{The name of the international marine bunkers flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$international_marine_bunkers}.}

\item{international_aviation_bunkers}{The name of the international aviation bunkers flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$international_aviation_bunkers}.}

\item{epsilon}{The name of the Epsilon matrix.
Default is "Epsilon".}
}
\value{

}
\description{
This function moves international bunkers flows to the Epsilon matrix.
The Epsilon balancing matrix is akin to an additional final demand matrix,
meaning that flows akin to final demand (i.e. where stocks increase) will be positive,
while flows akin to supply (i.e. where stocks decrease) will be negative.
}
\details{
See the Epsilon balacing matrix vignette for more information.
Note: one needs to add the column containing matrices names first,
most likely using the \code{IEATools::add_psut_matnames()} function.
}
