% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{non_energy_uses_to_balancing}
\alias{non_energy_uses_to_balancing}
\title{Relocates non-energy uses flows in the balancing matrix}
\usage{
non_energy_uses_to_balancing(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  e_dot = IEATools::iea_cols$e_dot,
  non_energy_uses = IEATools::tfc_flows$non_energy_use,
  balancing_matrix = "B"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which non-energy use flows need to be relocated in the Balancing matrix.}

\item{flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{non_energy_uses}{The string that identifies non-energy uses.
Default is \code{IEATools::tfc_flows$non_energy_use}.}

\item{balancing_matrix}{The name of the Balancing matrix.
Default is "B".}
}
\value{
A \code{.tidy_iea_df} with non-energy uses flows relocated in the Balancing matrix.
}
\description{
This function relocates non-energy uses to the balancing matrix B.
The Balancing matrix is akin to an additional final demand matrix.
}
\details{
See the Balancing matrix vignette for more information.
Note: one needs to add the column containing matrices names first,
most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
tibble::add_row(
 Country = "A",
 Method = "PCM",
 Energy.type = "E",
 Last.stage = "Final",
 Year = 2018,
 Ledger.side = "Supply",
 Flow.aggregation.point = "Consumption",
 Flow = "Non-energy use in whatever sector",
 Product = "Electricity",
 Unit = "ktoe",
 E.dot = 10,
) \%>\%
IEATools::add_psut_matnames() \%>\%
dplyr::filter(matnames == "B")
}
