% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_fossil_fuels}
\alias{specify_elect_heat_fossil_fuels}
\title{Specifies electricity and heat products}
\usage{
specify_elect_heat_fossil_fuels(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  eiou_flows = IEATools::aggregation_flows$energy_industry_own_use,
  main_act_prod_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_prod_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_prod_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  autoprod_elect = "Autoproducer electricity plants",
  autoprod_chp = "Autoproducer CHP plants",
  autoprod_heat = "Autoproducer heat plants",
  oil_products = "Oil products",
  coal_products = "Coal products",
  natural_gas = "Natural gas",
  other_products = "Other products",
  product_type = "Product type",
  share_inputs_from_Func = "Share_inputs_from_Func",
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{Name of the \code{.tidy_iea_df} for which electricity and heat products
need to be specified with origin product type.}

\item{country, method, energy_type, last_stage, year, ledger_side, flow_aggregation_point, flow, product, e_dot, unit}{See \code{IEATools::iea_cols}.}

\item{transformation_processes}{Name of transformation process flows in data frame.
Default is IEATools::aggregation_flows$transformation_processes.}

\item{eiou_flows}{Name of energy industry own use flows in data frame.
Default is IEATools::aggregation_flows$energy_industry_own_use.}

\item{main_act_prod_elect}{Name of electricity producing plants.
Default is IEATools::main_act_plants$main_act_prod_elect_plants.}

\item{main_act_prod_chp}{Name of CHP producing plants.
Default is IEATools::main_act_plants$main_act_prod_chp_plants}

\item{main_act_prod_heat}{Name of heat producing plants.
Default is IEATools::main_act_plants$main_act_prod_heat_plants}

\item{autoprod_elect}{Name of autoproducer electricity plants.
Default is "Autoproducer electricity plants".}

\item{autoprod_chp}{Name of autoproducer CHP plants.
Default is "Autoproducer CHP plants".}

\item{autoprod_heat}{Name of autoproducer heat plants.
Default is "Autoproducer heat plants".}

\item{oil_products}{Name of oil products. Default is "Oil products".}

\item{coal_products}{Name of coal products. Default is "Coal products".}

\item{natural_gas}{Name of natural gas. Default is "Natural gas".}

\item{other_products}{Name of other products. Default is "Other products".}

\item{product_type}{Name of a temporary column. Default is "Product type".}

\item{share_inputs_from_Func}{Name of a temporary column. Default is "Share_inputs_from_Func".}

\item{negzeropos}{Temporary column name. Default is ".netzeropos".}
}
\value{
Returns a \code{.tidy_iea_df} with electricity and heat products specified
with origin product type.
}
\description{
This function specified the electricity and heat energy products by the type of energy carrier
that has been used to produce them, at the moment "Oil products", "Coal products", "Natural gas", and "Other products".
To do so, the code keeps the main activity and autoproducer electricity, heat and CHP activities,
and calculates the inputs shares by each of the four product types group. The same shares are ascribe to electricity and heat ouputs.
New industries that are now specified as function of the product type they take as input.
}
\examples{
A_B_path <- system.file("extdata/A_B_data_full_2018_format.csv", package = "ECCTools")
IEATools::load_tidy_iea_df(A_B_path) \%>\%
IEATools::specify_all() \%>\%
specify_elect_heat_fossil_fuels()
}
