% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_markets}
\alias{specify_elect_heat_markets}
\title{Title}
\usage{
specify_elect_heat_markets(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{negzeropos}{}
}
\value{

}
\description{
Title
}
