% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_nuclear}
\alias{specify_elect_heat_nuclear}
\title{Specifies electricity and heat from nuclear}
\usage{
specify_elect_heat_nuclear(
  .tidy_iea_df,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which electricity and heat products coming from the nuclear industry need to be specified.}

\item{flow_aggregation_point, flow, e_dot, product}{See \code{IEATools::iea_cols}.}

\item{transformation_processes}{The name of transformation processes in the data frame.
Default is \code{IEATools::aggregation_flows$transformation_processes}.}

\item{nuclear_industry}{The name of the nuclear industry in the flow column of the data frame.
Default is \code{IEATools::eiou_flows$nuclear_industry}.}

\item{negzeropos}{Temporary column name. Default is ".netzeropos".}
}
\value{
A \code{.tidy_iea_df} with electricity and heat products specified when they come from the nuclear industry.
}
\description{
This function selects electricity and heat produced by the nuclear industry, and specifies them, so that the new product name is
respectively "Electricity \code{[}from Nuclear\verb{]}" and "Heat \code{[}from Nuclear\verb{]}".
}
