% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_renewables}
\alias{specify_elect_heat_renewables}
\title{Title}
\usage{
specify_elect_heat_renewables(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  eiou_flows = IEATools::aggregation_flows$energy_industry_own_use,
  main_act_prod_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_prod_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_prod_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  autoprod_elect = "Autoproducer electricity plants",
  autoprod_chp = "Autoproducer CHP plants",
  autoprod_heat = "Autoproducer heat plants",
  hydro = "Hydro",
  geothermal = "Geothermal",
  solar_pv = "Solar photovoltaics",
  solar_th = "Solar thermal",
  tide_wave_ocean = "Tide, wave and ocean",
  wind = "Wind",
  electricity = "Electricity",
  heat = "Heat",
  ratio_hydro_to_input = 1,
  ratio_solar_PV_to_input = 1,
  ratio_solar_thermal_elect_to_input = 0.33,
  ratio_solar_thermal_heat_to_input = 1,
  ratio_wind_to_input = 1,
  ratio_geothermal_elect_to_input = 0.1,
  ratio_geothermal_heat_to_input = 0.5,
  ratio_tidal_wave_to_input = 1,
  share_elect_output_From_Func = "Share_Output_Elect_From_Func",
  share_renewables_From_Func = "Share_Renewables_From_Func",
  e_dot_renewables = "E_dot_Renewables",
  e_dot_rest = "E_dot_Rest",
  negzeropos = ".netzeropos"
)
}
\arguments{
\item{negzeropos}{}
}
\value{

}
\description{
Title
}
