% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{specify_losses_as_industry}
\alias{specify_losses_as_industry}
\title{Specifies losses as a transformation industry}
\usage{
specify_losses_as_industry(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  losses = IEATools::tfc_compare_flows$losses,
  imports = IEATools::interface_industries$imports,
  stat_diffs = IEATools::tfc_compare_flows$statistical_differences,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  supply = IEATools::ledger_sides$supply,
  observation_string = "Observation_String",
  string_to_check = "String_to_Check"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy__iea_df} for which losses need to be specified as transformation industry.}

\item{country, method, energy_type, last_stage, ledger_side, flow_aggregation_point, year, flow, product, unit, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The name of the matrix name column in the input data frame.}

\item{losses}{The name of the "Losses" flows in the input data frame.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{imports}{The name of "Imports" flows in the input data frame.
Default is \code{IEATools::interface_industries$imports}.}

\item{stat_diffs}{The name of "Statistical differences" flows in the input data frame.
Default is \code{IEATools::tfc_compare_flows$statistical_differences}.}

\item{transformation_processes}{The name of transformation processes in the flow aggregation point column.
Default is \code{IEATools::tfc_compare_flows$transformation_processes}.}

\item{supply}{The name of the supply ledger side.
Default is \code{IEATools::ledger_sides$supply}.}

\item{observation_string}{The name of a temporary column identifying the observation.
Default is "Observation_String".}
}
\value{
A \code{.tidy_iea_df} with losses specified as a transformation industry.
}
\description{
Losses are originally treated as a final demand flow in the IEA WEEB balances. This function allows analysts to redefine losses
as a transformation industry, which takes as input "Product (before Losses)" and provides "Product".
Hence, final demand sectors still consume "Product", but where relevant, products go first through a "Losses (of Product)" industry.
}
\details{
TO DO:!!! Importantly, we need to check that values do not become negative...!!!
}
