% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_to_gma.R
\name{specify_MR_R}
\alias{specify_MR_R}
\title{Specifies Multi-Regional R matrix}
\usage{
specify_MR_R(
  .tidy_iea_df,
  R_matrix = IEATools::psut_cols$R,
  matnames = IEATools::mat_meta_cols$matnames,
  imports = IEATools::interface_industries$imports,
  country = IEATools::iea_cols$country,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  aggregate_country_name = "World"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} from which the Multi-Regional R matrix needs to be created.}

\item{R_matrix}{The name of the R matrix.
Default is \code{IEATools::psut_cols$R}.}

\item{matnames}{The name of the matrix names column.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{imports}{The name of the Imports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$imports}.}

\item{country, flow, product}{See \code{IEATools::iea_cols}.}

\item{aggregate_country_name}{The name of the new region that gathers all flows of the \code{.tidy_iea_df}.
Default is "World".}
}
\value{
A \code{.tidy_iea_df} representing the Multi-Regional Resources matrix, with all flows specified by country.
}
\description{
This function specified the Multi-Regional Resources matrix.
Only flows that belong to the Multi-Regional Resources matrix are returned,
other flows are filtered out.
}
\details{
The specification process is to modify both the flow and the product produced
so that the producing country is added as prefix using brackets first.
Note: matrix names need to be added, most likely using the \code{IEATools::add_psut_matnames}, prior to using the function.
}
\examples{
tidy_AB_data \%>\%
 IEATools::add_psut_matnames() \%>\%
 specify_MR_R() \%>\%
 print()
}
