% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{specify_MR_V}
\alias{specify_MR_V}
\title{Specifies Multi-Regional V matrix}
\usage{
specify_MR_V(
  .tidy_iea_df,
  V_matrix = IEATools::psut_cols$V,
  Epsilon_matrix = IEATools::psut_cols$epsilon,
  matnames = IEATools::mat_meta_cols$matnames,
  imports = IEATools::interface_industries$imports,
  country = IEATools::iea_cols$country,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  aggregate_country_name = "World"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} from which the Multi-Regional R matrix needs to be created.}

\item{V_matrix}{The name of the V matrix.
Default is \code{IEATools::psut_cols$V}.}

\item{Epsilon_matrix}{The name of the Epsilon matrix.
Default is \code{IEATools::psut_cols$epsilon}.}

\item{matnames}{The name of the matrix names column.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{imports}{The name of the Imports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$imports}.}

\item{country, flow, product, e_dot}{See \code{IEATools::iea_cols}.}

\item{aggregate_country_name}{The name of the new region that gathers all flows of the \code{.tidy_iea_df}.
Default is "World".}
}
\value{
A \code{.tidy_iea_df} representing the Multi-Regional Supply (V) matrix, with all flows specified by country.
}
\description{
This function specified the Multi-Regional Supply (V) matrix.
Only flows that belong to the Multi-Regional Supply matrix are returned,
other flows are filtered out. Exceptions are flows akin to supply (with E.dot < 0) but that have been
previously re-directed to the Epsilon matrix by the analyst - those are also kept.
}
\details{
The specification process is to modify both the flow and the product produced
so that the producing country is added as prefix using brackets first.
Note: matrix names need to be added, most likely using the \code{IEATools::add_psut_matnames}, prior to using the function.
}
\examples{
tidy_AB_data \%>\%
 IEATools::add_psut_matnames() \%>\%
 specify_MR_V() \%>\%
 print()
}
