% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{stat_diffs_to_balancing}
\alias{stat_diffs_to_balancing}
\title{Moving statistical differences flows to Balancing matrix}
\usage{
stat_diffs_to_balancing(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  e_dot = IEATools::iea_cols$e_dot,
  stat_diffs = "Statistical differences",
  ledger_side = IEATools::iea_cols$ledger_side,
  balancing_matrix = "B"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which statistical differences flows need to be sent to the Balancing matrix.}

\item{flow, e_dot, ledger_side}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{stat_diffs}{The name of the statistical differences flows.
Default is "Statistical differences".}

\item{balancing_matrix}{The name of the Balancing matrix.
Default is "B".}
}
\value{
A \code{.tidy_iea_df} for which statistical differences flows have been send to the Balancing matrix.
}
\description{
This function sends statistical differences flows to a balancing matrix Balancing.
The Balancing matrix is akin to an additional final demand matrix,
meaning that flows akin to final demand will be positive,
while flows akin to supply will be negative.
}
\details{
See the Balancing matrix vignette for more information.
Note: one needs to add the column containing matrices names first,
most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
tibble::add_row(
Country = "A",
Method = "PCM",
Energy.type = "Energy",
Last.stage = "Final",
Year = 2018,
Ledger.side = "Supply",
Flow.aggregation.point = "TFC compare",
Flow = "Statistical differences",
Product = "Crude oil",
Unit = "ktoe",
E.dot = 10
) \%>\%
IEATools::add_psut_matnames() \%>\%
stat_diffs_to_balancing() \%>\%
dplyr::filter(stringr::str_detect(Flow, "Statistical differences")) \%>\%
print()
}
