% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleans.R
\name{stat_diffs_to_epsilon}
\alias{stat_diffs_to_epsilon}
\title{Moving statistical differences flows to Epsilon matrix}
\usage{
stat_diffs_to_epsilon(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  matnames = "matnames",
  e_dot = IEATools::iea_cols$e_dot,
  stat_diffs = "Statistical differences",
  epsilon = "Epsilon"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which statistical differences flows need to be sent to the Epsilon matrix.}

\item{flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{stat_diffs}{The name of the statistical differences flows.
Default is "Statistical differences".}

\item{epsilon}{The name of the Epsilon matrix.
Default is "Epsilon".}
}
\value{
A \code{.tidy_iea_df} for which statistical differences flows have been send to the Epsilon matrix.
}
\description{
This function sends statistical differences flows to a balancing matrix Epsilon.
The Epsilon balancing matrix is akin to an additional final demand matrix,
meaning that flows akin to final demand will be positive,
while flows akin to supply will be negative.
}
\details{
See the Epsilon balacing matrix vignette for more information.
Note: one needs to add the column containing matrices names first,
most likely using the \code{IEATools::add_psut_matnames()} function.
}
\examples{
tidy_AB_data \%>\%
tibble::add_row(
Country = "A",
Method = "PCM",
Energy.type = "Energy",
Last.stage = "Final",
Year = 2018,
Ledger.side = "Supply",
Flow.aggregation.point = "TFC compare",
Flow = "Statistical differences",
Product = "Crude oil",
Unit = "ktoe",
E.dot = 10
) \%>\%
IEATools::add_psut_matnames() \%>\%
stat_diffs_to_epsilon() \%>\%
dplyr::filter(stringr::str_detect(Flow, "Statistical differences")) \%>\%
print()
}
