% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_AB_data}
\alias{tidy_AB_data}
\title{tidy_AB_data}
\format{
A \code{.tidy_iea_df} with 11 columns.
}
\usage{
tidy_AB_data
}
\description{
A \code{.tidy_iea_df} representing the energy conversion chain for both countries A and B.
These data are used as an example in many place in the package.
}
\examples{
tidy_AB_data
}
\keyword{datasets}
