% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_to_gma}
\alias{transform_to_gma}
\title{Transforms to Global Market Assumption}
\usage{
transform_to_gma(.tidy_iea_df)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} that needs to be converted into a data frame describing a Global Energy Conversion Chain,
adopting the Global Market Assumption.}
}
\value{
A \code{.tidy_iea_df} describin a Global Energy Conversion Chain, adopting a Global Market Perspective.
}
\description{
This function transforms a \code{.tidy_iea_df} that contains the representation of the Energy Conversion Chain for multiple
countries into a new tidy_iea_df that represents a Global Energy Conversion Chain, adopting the Global Market Assumption.
}
\details{
This function runs sequentially the following functions:
\itemize{
\item \code{specify_MR_R()};
\item \code{specify_MR_V()};
\item \code{specify_MR_Y_U_gma()};
and then binds all rows obtained using \code{dplyr::bind_rows()}.
}

Note: running this function to transform to the Global Market Assumption only makes sense when the country coverage is
global, or close to global (i.e. only countries consuming a very small fraction of global energy consumption, and only
producing a very small fraction of global energy producition, are missing).
}
\examples{
tidy_AB_data \%>\%
IEATools::add_psut_matnames() \%>\%
transform_to_gma()
}
