% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_fossil_fuels}
\alias{specify_elect_heat_fossil_fuels}
\title{Title}
\usage{
specify_elect_heat_fossil_fuels(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  eiou_flows = IEATools::aggregation_flows$energy_industry_own_use,
  main_act_prod_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_prod_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_prod_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  autoprod_elect = "Autoproducer electricity plants",
  autoprod_chp = "Autoproducer CHP plants",
  autoprod_heat = "Autoproducer heat plants",
  oil_products = "Oil products",
  coal_products = "Coal products",
  natural_gas = "Natural gas",
  other_products = "Other products",
  product_type = "Product type",
  share_inputs_from_Func = "Share_inputs_from_Func",
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{negzeropos}{}
}
\value{

}
\description{
Title
}
