% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_elect_heat.R
\name{specify_elect_heat_markets}
\alias{specify_elect_heat_markets}
\title{Specified electricity and heat markets}
\usage{
specify_elect_heat_markets(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which electricity and heat markets need to be specified.}

\item{country, method, energy_type, last_stage, year, ledger_side, flow_aggregation_point, flow, product, e_dot, unit}{See \code{IEATools::iea_cols}.}

\item{transformation_processes}{The name of transformation processes in the data frame.
Default is IEATools::aggregation_flows$transformation_processes.}

\item{negzeropos}{Temporary column name. Default is ".netzeropos".}
}
\value{
Returns a \code{.tidy_iea_df} with specified electricity and heat markets.
}
\description{
This function specifies electricity and heat markets. It does so by selecting production flows (V matrix) of:
"Electricity \code{[}from Oil products\verb{]}", "Electricity \code{[}from Coal products\verb{]}", "Electricity \code{[}from Natural gas\verb{]}",
"Electricity \code{[}from Other products\verb{]}", and "Electricity \code{[}from Renewables\verb{]}", and routing them as inputs to a new industry: "Electricity market".
The electricity market industry then produces "Electricity" in the same amount that it receives as input.
Exactly the same process is conducted for heat.
}
